Almost Wrapped Up
Needs style hanging but before messing with any of the style we need to fix claims not showing up on the admin side. There are some other little bugs but the data base should be working. I believe I have a claims card template but it might need to be completely scrapped and rewritten. On the claims side there should be a button to email the person and resolve claim and the claim is a summary of all the data on the person and item.