Poliishedish
Please leave a message of what i can change after you stress test it. I am tired so im gonna break for a bit if anyone wants to hop on call and work on it im willing to just to polish this up. We also need to get started on video script andd storyboarding.